Move smaps and smaps_rollup parsing into modules
This commit is code reorganization. I have not done anything to change
the mechanism here, just how it's used.
